+++ dependencies/glfw/CMakeLists.txt
@@ -2,8 +2,6 @@ include(FetchContent)
 
 FetchContent_Declare(glfw 
     URL https://github.com/glfw/glfw/releases/download/3.4/glfw-3.4.zip
-    URL_HASH SHA256=b5ec004b2712fd08e8861dc271428f048775200a2df719ccf575143ba749a3e9
-    DOWNLOAD_EXTRACT_TIMESTAMP true
 )
 
 FetchContent_MakeAvailable(glfw)
