+++ src/shared/CMakeLists.txt
@@ -1,9 +1,7 @@
 set(LIB_NAME game_shared_lib)
 
 set(FILES
-    game/servergamestate.cpp
     game/event.cpp
-    game/gamelogic/object.cpp
 
     network/session.cpp
