--- src/shared/game/gamelogic/creature.cpp
@@ -1,7 +1,0 @@
-#include "shared/game/gamelogic/creature.hpp"
-#include "shared/game/sharedgamestate/sharedobject.hpp"
-
-SharedObject Creature::generateSharedObject() {
-    auto so = Creature::generateSharedObject();
-    so.stats = this->stats;
-}
