--- src/shared/utilities/smartvector.cpp
@@ -1,4 +1,0 @@
-#include "shared/utilities/smartvector.hpp"
-
-//	TODO: See if smartvector.hpp can be refactored so that the implementation of
-//	SmartVector can be placed here.
